changes based on comments
